# swaggerDesign.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an api server for a portfolio builder website.\n",
    "version": "1.0.0",
    "title": "Portfolio Builder",
    "contact": {
      "email": "tshiyoyi@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "cse341-fab.herokuapp.com",
  "basePath": "/",
...
  "schemes": ["https"],
  "paths": {
...
    "/user/{username}": {
      "get": {
        "tags": ["user"],
        "summary": "Get user by user name",
        "operationId": "getUserByName",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be fetched.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["user"],
        "summary": "Updated user",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "name that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully modified user"
          },
          "400": {
            "description": "Invalid user supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "summary": "Delete user",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted user"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "_id": {
          "type": "integer",
          "format": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "info": {
          "$ref": "#/definitions/User_info"
        },
        "profile": {
          "$ref": "#/definitions/User_profile"
        }
      }
    },
...
    "User_info": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "currentLocation": {
          "type": "string"
        },
        "openToNewOpportunities": {
          "type": "boolean"
        },
        "profileIsPublic": {
          "type": "boolean"
        },
        "theme_name": {
          "type": "string"
        }
      }
    },
    "User_profile_experience": {
      "type": "object",
      "properties": {
        "jobTitle": {
          "type": "string"
        },
        "jobDescription": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "dateStart": {
          "type": "string"
        },
        "dateEnd": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "location": {
          "type": "string"
        }
      }
    },
    "User_profile_education": {
      "type": "object",
      "properties": {
        "institution": {
          "type": "string"
        },
        "programTitle": {
          "type": "string"
        },
        "programDescription": {
          "type": "string"
        },
        "dateStart": {
          "type": "string"
        },
        "dateEnd": {
          "type": "string"
        },
        "completed": {
          "type": "boolean"
        },
        "location": {
          "type": "string"
        },
        "topics": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "User_profile_projects": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "dateStart": {
          "type": "string"
        },
        "dateEnd": {
          "type": "string"
        },
        "githubUrl": {
          "type": "string"
        },
        "hostedUrl": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "User_profile_references": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "relationship": {
          "type": "string"
        },
        "contactInfo": {
          "type": "string"
        },
        "yearsKnown": {
          "type": "integer",
          "format": "int32"
        },
        "knownFromWhere": {
          "type": "string"
        }
      }
    },
    "User_profile": {
      "type": "object",
      "properties": {
        "experience": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User_profile_experience"
          }
        },
        "education": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User_profile_education"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User_profile_projects"
          }
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "references": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User_profile_references"
          }
        }
      }
    }
  }
}